moving motor_imagery into the analytics folder
