+++ brainsquared/analytics/motor_imagery.py
